chore: save stable todayhabit state
