+++ src/components/ui/Header/Header.jsx
@@ -1,16 +1,11 @@
-import styles from "./Header.module.css";
-import logo from "@/assets/logo/logo.svg";
+import styles from './Header.module.css';
+import logo from '@/assets/logo/logo.svg';
 
 export const Header = () => {
   return (
     <header className={styles.container}>
       <div className={styles.inner}>
         <a href="/">
-          <img
-            src={logo}
-            alt="TodayHabit"
-            className={styles.logo}
-          />
           <img src={logo} alt="TodayHabit" className={styles.logo} />
         </a>
       </div>
