+++ src/components/ui/PageCard/PageCard.module.css
@@ -1,6 +1,4 @@
 .container {
-  max-width: 960px;
-  margin: 0 auto;
   padding: 32px;
   border: 1px solid var(--color-gray-300);
   border-radius: var(--spacing-md);
