+++ src/components/ui/PageHeader/PageHeader.module.css
@@ -1,9 +1,6 @@
 .container {
   display: flex;
   justify-content: space-between;
-  align-items: flex-start;
-  max-width: 960px;
-  margin: 0 auto;
 }
 
 .left {
@@ -67,4 +64,28 @@
   border: 1px solid var(--color-gray-200);
   background: #ffffff;
   font-size: var(--font-size-md);
+}
+.currentTime {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: var(--spacing-sm);
+}
+
+.timeLabel {
+  font-size: var(--font-size-md);
+  color: var(--color-text-primary);
+}
+
+.timeValue {
+  font-size: var(--font-size-md);
+  color: var(--color-gray-900);
+  border: 1px solid var(--color-gray-300);
+  border-radius: var(--spacing-md);
+  padding: var(--padding-sm);
+}
+
+.cardMeta {
+  font-size: var(--font-size-md);
+  opacity: 0.6;
 }
