+++ src/pages/RootLayout/RootLayout.jsx
@@ -1,12 +1,13 @@
 import { Outlet } from 'react-router';
 import { Header } from '@/components/ui/Header/Header';
+import styles from './RootLayout.module.css';
 
 export const RootLayout = () => {
   return (
-    <div className="pageBackground">
-      <div className="pageLayout">
+    <div className={styles.pageBackground}>
+      <div className={styles.pageLayout}>
         <Header />
-        <main className="pageContent">
+        <main className={styles.pageContent}>
           <Outlet />
         </main>
       </div>
