+++ src/pages/TodayHabitPage/TodayHabitPage.module.css
@@ -2,35 +2,9 @@
   padding: var(--padding-md) 0;
 }
 
-/* 현재시간 */
-.currentTime {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: var(--spacing-sm);
-}
-
-.timeLabel {
-  font-size: var(--font-size-md);
-  color: var(--color-text-primary);
-}
-
-.timeValue {
-  font-size: var(--font-size-md);
-  color: var(--color-gray-900);
-  border: 1px solid var(--color-gray-300);
-  border-radius: var(--spacing-md);
-  padding: var(--padding-sm);
-}
-
-.cardMeta {
-  font-size: var(--font-size-md);
-  opacity: 0.6;
-}
-
+/* card */
 .cardHeader {
   display: flex;
-  align-items: flex-end;
   gap: var(--spacing-lg);
   justify-content: center;
 }
