+++ src/styles/index.css
@@ -1,26 +1,7 @@
-* {
-  box-sizing: border-box;
+html,
+body,
+#root {
+  height: 100%;
+  margin: 0;
 }
 
-body {
-  font-family: 'Pretendard', sans-serif;
-}
-
-.pageBackground {
-  min-height: 100vh;
-  background-color: var(--color-background);
-}
-
-.pageLayout {
-  padding: 0 var(--page-gutter);
-  padding-top: 39px;
-}
-
-.pageContent {
-  background: var(--color-white);
-  max-width: 960px;
-  margin: 0 auto;
-  border-radius: var(--radius-md);
-  overflow: hidden;
-  padding: var(--padding-lg);
-}
